+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+## 0.3.0 - 2023-09-01
+### Added
+- Macro `ROTATE_LOCAL_MAP` that rotates the local map in order to keep x-axis in the direction of the car.
+- Subscriber on `/path` that enables checking collisions on the Path. Obstacles are published to `/path/obstacles`.
+- Parameter `~filter_map` that enables filtering of map from the LaserScan data.
+
+### Changed
+- ROS1 version is using `tf2` package.
+- `tf2` dependency is shared for both ROS versions.
+- Every cell on the static map with probability > 50 is treated as obstacle.
+- Node starts after first valid transformation is found.
+
+### Removed
+- `tf` dependency for ROS1.
+
 ## 0.2.0 - 2023-06-13
 ### Added
 - Publish local map to `/map/local`.